Move extract_image to utils module
