+++ tests/test_utils.py
@@ -1,3 +1,5 @@
+from unittest.mock import patch, mock_open
+
 import pytest
 
 from rpm_lockfile import utils
@@ -33,3 +35,21 @@ def test_find_containerfile(tmpdir, files, expected):
         assert actual == tmpdir / expected
     else:
         assert actual is None
+
+
+@pytest.mark.parametrize(
+    "file,expected",
+    [
+        ("""FROM registry.io/repository/base
+RUN something
+""", "registry.io/repository/base"),
+        ("""FROM registry.io/repository/build as build
+RUN build
+FROM registry.io/repository/base
+COPY --from=build /artifact /
+""", "registry.io/repository/base"),
+    ]
+)
+def test_extract_image(file, expected):
+    with patch("builtins.open", mock_open(read_data=file)):
+        assert utils.extract_image(file) == expected
